#184752104 A buyer should be able to add a product to wish list -merged

+++ database/models/cart.js
@@ -19,7 +19,7 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true
     },
     productId: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(36),
       allowNull: false
     },
     userId: {
